# docs/openapi.yaml
openapi: 3.0.3
info:
  title: StreamSphere Go API
  version: 1.0.0
  description: Contratos principales expuestos por el API Gateway.
servers:
  - url: http://localhost:8080
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    VideoInput:
      type: object
      required: [channelId, title]
      properties:
        channelId: { type: string, format: uuid }
        title: { type: string, minLength: 3, maxLength: 150 }
        description: { type: string }
        category: { type: string }
        tags:
          type: array
          items: { type: string }
        visibility:
          type: string
          enum: [PUBLIC, PRIVATE, UNLISTED]
paths:
...
  /api/videos:
    post:
      summary: Crear metadatos de video
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/VideoInput' }
      responses:
        '201': { description: Video creado en DRAFT }
  /api/videos/search:
    get:
      summary: Buscar videos públicos y publicados
      parameters:
        - { in: query, name: q, schema: { type: string } }
        - { in: query, name: category, schema: { type: string } }
        - { in: query, name: page, schema: { type: integer, default: 1 } }
        - { in: query, name: size, schema: { type: integer, default: 12 } }
      responses:
        '200': { description: Resultados paginados }
  /api/videos/{id}:
    get:
      summary: Consultar detalle de video
      parameters:
        - { in: path, name: id, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Detalle del video }
        '404': { description: Video inexistente o privado }
...